% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{dbGaP data use downloader}
\usage{
getStats()
}
\description{
This function simply queries the NCBI dbGaP data use
summary endpoint to generate a data.frame for local
use. The function uses queries to download the most
fine-grained data possible with the expectation that
the downloaded data will be processed in R for summaries,
etc.
}
